backend/requirements.txt
Flask>=2.0
Flask-Cors>=3.0
python-dotenv>=1.0
requests>=2.0
nltk>=3.0
